+++ SpatialGDK/Source/Private/Interop/SpatialStaticComponentView.cpp
@@ -3,6 +3,7 @@
 #include "Interop/SpatialStaticComponentView.h"
 
 #include "Schema/Component.h"
+#include "Schema/Interest.h"
 #include "Schema/Rotation.h"
 #include "Schema/Singleton.h"
