+++ SpatialGDK/Source/Public/EngineClasses/SpatialNetDriver.h
@@ -30,7 +30,6 @@ class USpatialStaticComponentView;
 class USnapshotManager;
 
 class UEntityRegistry;
-class UActorProxyRegistry;
 
 DECLARE_LOG_CATEGORY_EXTERN(LogSpatialOSNetDriver, Log, All);
 
@@ -121,8 +120,6 @@ public:
 	UEntityRegistry* EntityRegistry;
 	UPROPERTY()
 	USnapshotManager* SnapshotManager;
-	UPROPERTY()
-	UActorProxyRegistry* ActorProxyRegistry;
 
 	TMap<UClass*, TPair<AActor*, USpatialActorChannel*>> SingletonActorChannels;
